new_key_app/requirements.txt
Django
djangorestframework
djangorestframework-api-key
pytz
sqlparse
loguru
djangorestframework-simplejwt
